+++ car-management-system-api/src/main/java/com/naruhin/api/controller/ManufacturerRestController.java
@@ -1,6 +1,6 @@
-package com.naruhin.springbootexamplehillelhw5.web;
+package com.naruhin.api.controller;
 
-import com.naruhin.springbootexamplehillelhw5.dto.ManufacturerDTO;
+import com.naruhin.api.dto.ManufacturerDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
